+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.5.0] - 2022-09-20
+
 ### Added
 - (CYBRSOURCE-46) Optional order reference suffix
 - Functions to support Payer Auth